trigger help display using HelpException
