+++ src/dotnet/ParseResultExtensions.cs
@@ -7,18 +7,20 @@ using Microsoft.DotNet.Cli.CommandLine;
 
 namespace Microsoft.DotNet.Cli
 {
-    public static class ParserExtensions
-    {
-        public static ParseResult ParseFrom(
-            this CommandLine.Parser parser,
-            string context,
-            string[] args) =>
-            parser.Parse(context.Split(' ').Concat(args).ToArray());
-    }
-
     public static class ParseResultExtensions
     {
         public static void ShowHelp(this ParseResult parseResult) =>
             Console.WriteLine(parseResult.Command().HelpView());
+
+        public static void ShowHelpIfRequested(this ParseResult parseResult)
+        {
+            if (parseResult.HasOption("help"))
+            {
+
+                // NOTE: this is a temporary stage in refactoring toward the ClicCommandLineParser being used at the CLI entry point. 
+
+                throw new HelpException(parseResult.Command().HelpView());
+            }
+        }
     }
 }
