+++ src/dotnet/commands/dotnet-publish/Program.cs
@@ -25,6 +25,8 @@ namespace Microsoft.DotNet.Tools.Publish
 
             var result = parser.ParseFrom("dotnet publish", args);
 
+            result.ShowHelpIfRequested();
+
             msbuildArgs.Add("/t:Publish");
 
             var appliedPublishOption = result["dotnet"]["publish"];
